+++ sparv/core/run_snake.py
@@ -139,10 +139,10 @@ if not use_preloader:
         # something went wrong.
         exit_with_error_message(e.message, "sparv.modules." + module_name)
     except Exception as e:
-        errmsg = f"An error occurred while executing {module_name}:{f_name}:"
+        errmsg = f"An error occurred while executing {module_name}:{f_name}:\n\n  {type(e).__name__}: {e}"
         if logger.level > logging.DEBUG:
-            errmsg += f"\n\n  {type(e).__name__}: {e}\n\n" \
-                      "To display further details when errors occur, run Sparv with the '--log debug' argument."
+            errmsg += f"\n\nTo display further details when errors occur, run Sparv with the '--log debug' or " \
+                      "'--log-to-file debug' arguments."
         logger.error(errmsg)
         logger.debug(traceback.format_exc())
         sys.exit(123)